♻️(backend) split api files
As we aproach near 1000 lines in api.py, we have to split it in multiple
files.
